improve Opyn dashboard
